% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_post.R
\name{new_post}
\alias{new_post}
\title{Create a new Quarto post with pre-filled YAML}
\usage{
new_post(
  title,
  file = "index.qmd",
  description = "",
  author = "Mitch Henderson",
  date = Sys.Date(),
  draft = FALSE,
  title_limit = 40,
  open_file = TRUE
)
}
\arguments{
\item{title}{Name of the post}

\item{file}{Name of the .qmd file generated}

\item{description}{Post description}

\item{author}{Author name (default is "Mitch Henderson")}

\item{date}{Date of post (default is today)}

\item{draft}{Default is FALSE but when TRUE the post will not be rendered due to "_" prefix added}

\item{title_limit}{Number of characters beyond which a warning is provided}

\item{open_file}{Default is TRUE meaning the .qmd file is automatically opened.}
}
\value{
Nothing. Side effects include a newly created folder and file.
}
\description{
Create a new Quarto post with pre-filled YAML
}
\examples{
\dontrun{
new_post(
  title = "My new post",
  description = "My new post is about xyz",
  draft = TRUE
)
}
}
