% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_caption.R
\name{social_caption}
\alias{social_caption}
\title{Create a ggplot caption with social media details}
\usage{
social_caption(
  linkedin_name = "Mitch Henderson",
  bluesky_username = "mitchhenderson",
  github_username = "mitchhenderson",
  font_family = "Source Sans 3",
  icon_font_family = "Font Awesome 7 Brands",
  icon_colour = "#134A8E",
  font_colour = "grey20"
)
}
\arguments{
\item{linkedin_name}{Name on LinkedIn}

\item{bluesky_username}{Bluesky username}

\item{github_username}{GitHub username}

\item{font_family}{Font to use for social media handles}

\item{icon_font_family}{Font to use for the icons}

\item{icon_colour}{Hex colour code for icons}

\item{font_colour}{Hex colour code for social media handles}
}
\value{
A character vector
}
\description{
Create a ggplot caption with social media details
}
\examples{
socials <- social_caption(
  linkedin_name = "Mitch Henderson",
  bluesky_username = "mitchhenderson",
  github_username = "mitchhenderson"
  )
}
